+++ changelog.md
@@ -3,6 +3,7 @@
 * LLT-3777: Apply rebinding logic to all Apple platforms
 * LLT-4002: Review log-levels of spammy log messages
 * LLT-3990: Add initial heartbeat interval to feature config
+* LLT-4008: Add support for AAAA records in telio-dns
 
 <br>